Create accessible_comic.html
I think this comic is great and would love to share with my community. I created an accessible transcript for the comic. I also submitted a pull request for the transcript to the DP-3T repo. If you prefer not to accept my updates, that's totally fine, I would be happy to host an accessible version on my website. I took some liberties with a character name and descriptions, so please feel free to update my transcript if you prefer different phrasing.

Additionally, I added my name as author of the transcripts so that you don't have to take responsibility for critiques, mistakes, etc. If you prefer that my work remains anonymous, at least on your website, feel free to remove my name.